Revert "Split gui and anttask out of main binary (#746)"
This reverts commit a0fe4ed68296ead2dee8ba6a7c4bbfe48a443b50

Causes failure of the regression tests
```
Test [src/test/cases/unicode-space/UnicodeWhiteSpace1_5_0.test] finished successfully.
2371
Test failed, unexpected output: Difference in line 117254
2372
expected :[line: 2027 col: 33 match: -- --]
2373
but got  :[line: 2027 col: 33 match: --[--]
2374
Test output: line: 1 col: 1 match: --/* The following code was generated by JFlex 1.4.1 on 5/26/08 10:22 AM */--
```
