+++ jflex/src/main/java/jflex/BUILD.bazel
@@ -3,21 +3,9 @@ package(default_visibility = ["//visibility:public"])
 # cli, gui, and ant task
 java_library(
     name = "jflex_ui",
-    srcs = ["Main.java"],
+    srcs = ["Main.java"],  # cli
     deps = [
-        ":jflex_cli",
         "//jflex/src/main/java/jflex/anttask",
-        "//jflex/src/main/java/jflex/exceptions",
-        "//jflex/src/main/java/jflex/gui",
-        "@maven//:com_google_guava_guava",
-    ],
-)
-
-# cli only
-java_library(
-    name = "jflex_cli",
-    srcs = ["Cli.java"],  # cli
-    deps = [
         "//jflex/src/main/java/jflex/base",
         "//jflex/src/main/java/jflex/core",
         "//jflex/src/main/java/jflex/core/unicode",
