+++ testsuite/testcases/src/test/cases/java/java-1.input
@@ -2024,7 +2024,7 @@ public class Scanner implements sym {
    * @param argv   the command line, contains the filenames to run
    *               the scanner on.
    */
-  public static void main(String[] argv) {
+  public static void main(String argv[]) {
     if (argv.length == 0) {
       System.out.println("Usage : java Scanner <inputfile>");
     }
